requirements.txt
# Minimal requirements for Sanjaya Autopilot Platform
# For local development

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pyyaml>=6.0
gitpython>=3.1.0

# LLM providers (optional - install based on provider choice)
openai>=1.0.0
anthropic>=0.18.0

